[hg] add additional fallback locations to try applying diffs
If we fail to successfully apply a phabricator diff anywhere else, try
"master", "@", and ".".  In particular, this is needed in case the base
commit doesn't exist, and all of the files in the diff are new.  If all
of the files are new then we won't find any candidate commits from the
affected paths.