[hg] refactor arcanist ChangeSet handling
Refactor the logic to look at the change set type, rather than trying to
use old_path and current_path.  Arcanist doesn't always use old_path to
mean the path name in the old commit and current_path to mean the path
in the new revision.  For instance, in "MOVE_AWAY" changes, current_path
has the name of the path in the old revision (the location the file was
moved from), and old_path is None.