[hg] reduce the number of candidate base revisions searched
My repositories tend to grow lots of heads over time (for instance, all
of the applied diffs).  Just computing the list of revisions to search was
taking a long time.

To reduce the time taken here, only search master, plus heads with the
same author as the new diff being applied.