[hg] fix an issue creating commits with non-ASCII characters
Previously we were passing the author name into commitctx() as unicode.
When this was joined with the commit description this would fail if the
description was raw bytes with non-ASCII characters.

This makes sure we pass all fields in as bytes, which avoids trying to
decode any of the fields into unicode characters.