[arc] drop built-in Phabricator support, move logic to config
Remove the built-in logic for downloading and applying Phabricator
diffs.  Instead we now support loading a config module located at
~/.scmreview, and this config can supply a function for downloading and
applying diffs.

This will allow users to provide their own logic, which can be more
customized for their site, and can be customized on a per-repo basis.
This does make require more work on the users' part to define a custom
function to do this, however.  In the future we could add back some more
building blocks to help make this easier.