[eden] improve handling of revset arguments
Set `HGPLAIN=revsetalias` in the environment to allow expanding
user-defined revsetalias configs.

Also attempt to wrap integers in the revset input in `id()` to force
them to be interpreted as commit hashes rather than revnum values.
revnum values are a legacy Eden SCM behavior, and are disabled by
default, but are currently re-enabled when `HGPLAIN=1` is used.
There currently isn't a better way to disable revnum resolution.